Move RoomSelector.vue to src/components
